+++ ui/app/src/components/FolderMoveAlert/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2007-2022 Crafter Software Corporation. All Rights Reserved.
+ * Copyright (C) 2007-2025 Crafter Software Corporation. All Rights Reserved.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 3 as published by
@@ -14,9 +14,6 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export interface CMISItem {
-	itemId: string;
-	itemName: string;
-	mimeType: string;
-	size: number;
-}
+export { default } from './FolderMoveAlert';
+
+export * from './FolderMoveAlert';