+++ ui/app/src/state/reducers/dialogs/index.ts
@@ -50,6 +50,7 @@ import bulkCancelPackage from './bulkCancelPackage';
 import publishingPackageResubmit from './publishingPackageResubmit';
 import packageDetails from './packageDetails';
 import viewPackages from './viewPackages';
+import folderMoveAlert from './folderMoveAlertDialog';
 
 export default combineReducers({
 	confirm,
@@ -86,5 +87,6 @@ export default combineReducers({
 	cancelPackage,
 	bulkCancelPackage,
 	packageDetails,
-	viewPackages
+	viewPackages,
+	folderMoveAlert
 });